requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
requests==2.31.0
google-cloud-bigquery==3.15.0
db-dtypes==1.2.0

gunicorn
